requirements.txt
colorama==0.3.7
Django==1.10.4
djangorestframework==3.5.3
docopt==0.6.2
image==1.5.5
logilab-common==1.3.0
nltk==3.2.1
nose==1.3.7
numpy==1.11.3
Pillow==3.4.2
pipreqs==0.4.6
requests==2.12.4
six==1.10.0
yarg==0.1.9
